backend/requirements.txt
asgiref==3.8.1
autopep8==2.3.1
Django==5.1.4
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
setuptools>=75.0.0
pycodestyle==2.12.1
PyJWT==2.10.1
pytz==2024.2
sqlparse==0.5.2
toml==0.10.2
tomli==2.1.0
tzdata==2024.2
# Interview System Dependencies
python-dotenv==1.0.0
twilio==8.12.0
requests==2.31.0
# Website Analysis Dependencies
openai==0.28.1
beautifulsoup4==4.12.2
python-dateutil==2.8.2
